% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{admin_agg}
\alias{admin_agg}
\title{Aggregate variables at the specified administrative unit}
\usage{
admin_agg(
  data,
  context,
  context_AP,
  agg_level = NULL,
  data_name,
  WSC_AP,
  weights = NULL,
  data_sheet_name = NULL
)
}
\arguments{
\item{data}{data.frame containing the data to be aggregated}

\item{context}{character string identifying the context to be used in the function call.
This is to be used if multiple context (geographical or temporal) are being
analysed. For instance, if data is used for Burkina Faso in 2020 and 2019,
this column can help distinguish the indicators.}

\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example \href{https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing}{here} or in \code{WSCprocessing::context_AP}.}

\item{agg_level}{character string specifying which column should be used to
aggregate the data. This is is typically an administrative unit (e.g. province,
region, departement, admin2, etc.)}

\item{data_name}{character string identifying the name of the data frame used
in \code{data}. Should be equivalent to the \code{data_source_name} called in \code{context_AP}.}

\item{WSC_AP}{data.frame with the general WSC analysis plan (AP) than can be found \href{https://docs.google.com/spreadsheets/d/1TKxD_DyBTTN6onxYiooqtcI_TVSwPfeE-t7ZHK1zzMU/edit?usp=sharing}{here} or as an object in the package (\code{WSCprocessing::WSC_AP})}

\item{weights}{string containing the column name where weights are stored.}

\item{data_sheet_name}{character string with the name of the \code{sheet_name}}
}
\value{
data.frame containing the aggregated data according to context_AP and WSC_AP
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\dontrun{
agg_admin2 <- admin_agg(data = WSCprocessing::bfa_msna_2020, context ="bfa_2020",
context_AP = WSCprocessing::context_AP,  agg_level = "admin2", data_name = "bfa_msna_2020",
weights = "weights_sampling",data_sheet_name = "BFA_MSNA_2020_dataset_cleanedWeighted_ADM1",
WSC_AP = WSCprocessing::WSC_AP)
}
}
