% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_country}
\alias{analyse_country}
\title{Analyse data according to specified analysis plan}
\usage{
analyse_country(
  context_AP,
  admin_analysis,
  country_iso3,
  pop_df = NULL,
  google_api_key = NULL,
  pcodes_df = NULL,
  ...
)
}
\arguments{
\item{context_AP}{data.frame with context specific analysis plan (AP) that
links the indicators in the WSC AP to the datasets used in the context
analysis. See an example
\href{https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing}{here}
or in \code{\link{context_AP}}.}

\item{admin_analysis}{character string indicating at which administrative
unit the analysis should be aggregated. Should follow the synthax: admin +
admin_unit_number (e.g. admin1, admin2 , etc.)}

\item{country_iso3}{character string with the ISO3 country code for the
country analysed. See
{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{here} for more details.}

\item{pop_df}{data.frame containing population figures for each admin unit
being analysed. This data frame is used to weight the datasets when
aggregating from one administrative unit to another. The data frame must
contain at a minimum: \describe{ \item{admin1}{contains names of
administrative unit 1: regions, provinces, etc.} \item{admin2}{contains
names of administrative unit 2: districts, departments, etc.}
\item{total_pop}{contains the population figures associated with each
administrative unit} } If additional administrative units are required, the
name of the column should follow this structure: admin+number (e.g.
admin3). If left to the default (NULL), the package will try to download a
dataframe from \href{https://data.humdata.org}{HDX} Common Operational Datasets
(COD). This is a quite hazardous approach as HDX COD's are not formatted in
a consistent way.}

\item{google_api_key}{a string with a google API key. This is used to recode
administrative units. See vignette("googleway-vignette") for more details.}

\item{pcodes_df}{dataframe containing matching key between pcodes and admin
names. For more details see
\href{https://github.com/elliottmess/pcodesOCHA}{here}.}

\item{...}{parameters to be passed to other functions, must be following the
format of a named list (e.g. WSC_AP = this_df).}
}
\value{
data.frame containing the analysed data
}
\description{
This function takes an analysis plan following the \code{\link{context_AP}}
structure, a specified administrative level of analysis ("admin1","admin2",
or "admin3"), and a country iso3 code to conduct the analysis on.
Additional parameters are also recommended, especially \code{pop_df}, which
is used to weight data when aggregated to a higher administrative unit. If it
is not provided, the script will try to find a dataset on
\href{https://data.humdata.org}{HDX} which is an hasardous procedure.
}
\details{
This function differs from \code{\link[=analyse_DAP]{analyse_DAP()}} by combining all the data sources
provided in analysis plan, cleaning administrative names over all data
sources, imputing data available at a lower or higher administrative unit
than the one chosen and \strong{breaking ties} based on the data source reliability if
there is more than one data source for an indicator.

The imputation of the data from one administrative unit to another level is
done by imputing the value value from a higher administrative unit to a lower
one, and by imputing a weighted by population value from a lower one to a
higher one. For instance if our \code{admin_analysis} is "admin2", all data
sources available at admin1 only will be imputed as is (all admin2 receive
the same value as their respective admin1), using
\code{assign_result_high2low()}. All data sources available at
admin3 will be imputed using a weighted mean based on the population figures
provided by \code{pop_df} to the admin2 using \code{assign_result_low2high()}.
}
\examples{
\dontrun{
result <- analyse_country(context_AP, admin_analysis = "admin2", country_iso3 = "BFA")
}
}
\seealso{
Other analysis functions: 
\code{\link{analyse_DAP}()},
\code{\link{analyse_data}()},
\code{\link{analyse_source_all_sheets}()},
\code{\link{analyse_var}()}
}
\concept{analysis functions}
