% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse (conflicted).R, R/analyse.R
\name{analyse_DAP}
\alias{analyse_DAP}
\title{Analyse data according to the contextual data analysis plan (DAP)}
\usage{
analyse_DAP(
  context_AP,
  country_iso3,
  google_api_key = NULL,
  pcodes_df = NULL,
  ...
)

analyse_DAP(
  context_AP,
  country_iso3,
  google_api_key = NULL,
  pcodes_df = NULL,
  ...
)
}
\arguments{
\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example [here](https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing) or in ```WSCprocessing::context_AP```.}

\item{country_iso3}{character string with the ISO3 country code for the
country analysed. See [here](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3) for more details.}

\item{google_api_key}{a string with a google API key. See vignette("googleway-vignette") for more details.}

\item{pcodes_df}{dataframe containing matching key between pcodes and admin
names. For more details see [here](https://github.com/elliottmess/pcodesOCHA).}

\item{...}{parameters to be passed to other functions, must be following the
format of a named list (e.g. WSC_AP = this_df).}
}
\value{
data.frame containing the analysed data

data.frame containing the analysed data
}
\description{
Analyse data according to the contextual data analysis plan (DAP)

Analyse data according to the contextual data analysis plan (DAP)
}
\examples{
\dontrun{
context_AP <- WSCprocessing::context_AP

result <- analyse_DAP(context_AP, country_iso3 = "BFA")

}
\dontrun{
context_AP <- WSCprocessing::context_AP

result <- analyse_DAP(context_AP, country_iso3 = "BFA")

}
}
