% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{analyse_source}
\alias{analyse_source}
\title{Analyse data source}
\usage{
analyse_source(source_name, context_AP)
}
\arguments{
\item{source_name}{character string containing the name of the data source to
be analysed according to the analysis plan.}

\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example [here](https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing) or in ```WSCprocessing::context_AP```.}
}
\value{
data.frame containing the analysed data
}
\description{
Function to analyse data source according to the analysis plan provided.
}
\examples{
\dontrun{
analysed_source <- analyse_source("smart_rapid_idp_2020", context_AP)
}
}
