% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse (conflicted).R, R/analyse.R
\name{analyse_source_all_sheets}
\alias{analyse_source_all_sheets}
\title{Analyse all data sheets in a data source
name = unlist(data_AP$indicator_code_source)
Function to analyse data source according to the analysis plan provided.}
\usage{
analyse_source_all_sheets(
  source_name,
  context_AP,
  WSC_AP = WSCprocessing::WSC_AP,
  ...
)

analyse_source_all_sheets(
  source_name,
  context_AP,
  WSC_AP = WSCprocessing::WSC_AP,
  ...
)
}
\arguments{
\item{source_name}{character string containing the name of the data source to
be analysed according to the analysis plan.}

\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example [here](https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing) or in ```WSCprocessing::context_AP```.}

\item{WSC_AP}{data.frame with the general WSC analysis plan (AP) than can be
found \href{https://docs.google.com/spreadsheets/d/1TKxD_DyBTTN6onxYiooqtcI_TVSwPfeE-t7ZHK1zzMU/edit?usp=sharing}{here} or as an object in the package (```WSCprocessing::WSC_AP```)}

\item{...}{parameters to be passed to other functions, must be following the
format of a named list (e.g. WSC_AP = this_df).}
}
\value{
data.frame containing the analysed data

data.frame containing the analysed data
}
\description{
Analyse all data sheets in a data source
name = unlist(data_AP$indicator_code_source)
Function to analyse data source according to the analysis plan provided.

Analyse all data sheets in a data source
name = unlist(data_AP$indicator_code_source)
Function to analyse data source according to the analysis plan provided.
}
\examples{
\dontrun{
analysed_source <- analyse_source_all_sheets("smart_rapid_idp_2020", WSCprocessing::context_AP)
}
\dontrun{
analysed_source <- analyse_source_all_sheets("smart_rapid_idp_2020", WSCprocessing::context_AP)
}
}
