% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_var}
\alias{analyse_var}
\title{Analyse a given variable}
\usage{
analyse_var(design_data, var_analyse, agg_level, ...)
}
\arguments{
\item{design_data}{\code{\link[srvyr]{as_survey_design}} object containing the data,
weights, cluster IDs, and other elements required for complexe surveys
analysis.}

\item{var_analyse}{character string identifying the column with the variable
to analyse.}

\item{agg_level}{character string identifying the column to be used to
aggregate the data.}

\item{source}{character string containing the name of the data source
analysed.}
}
\value{
dataframe containing the analysed data. The returned table follows
this structure:
\describe{
\item{agg_level}{contains the administrative unit at which the data
has been aggregated. The name of the column will be the value of the
agg_level argument provided.}
\item{indicator}{Identifier of the indicator (variable) being analysed.
The name of the column will be the value of the var_analyse argument
provided.}
\item{choice}{contains the value }
}
}
\description{
Analyse (summarise/aggregate) a variable contained in a
\code{\link[srvyr]{as_survey_design}} object by getting the weighted mean of
the data if it is a numerical  variable, or the weighted mean frequency of
its factors if it is a  character/factor variable.
}
\examples{

\dontrun{

this_dataAP <- get_dataAP("REACH-MSNA-2020", NA, context_AP)

df <- bfa_msna_2020

this_design <- srvyr::as_survey(df ,
ids = sampling_id,
weights = weights_sampling)

var_analyse <- "dispo_lave_main"

result <- analyse_var(this_design, var_analyse,
agg_level = this_dataAP$admin_level)
}
}
\seealso{
Other analysis functions: 
\code{\link{analyse_DAP}()},
\code{\link{analyse_country}()},
\code{\link{analyse_data}()},
\code{\link{analyse_source_all_sheets}()}
}
\concept{analysis functions}
