% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high2low.R
\name{assign_result_high2low}
\alias{assign_result_high2low}
\title{Assign higher administrative unit data to a lower one}
\usage{
assign_result_high2low(high_df, high_df_name, low_admin, pop_df, context_AP)
}
\arguments{
\item{high_df}{data.frame with the higher administrative unit data to be
assigned.}

\item{high_df_name}{character string with the name of the
\code{high_df}. Must follow the format:
     \code{data_source_name} - \code{data_sheet_name} columns of
\code{context_AP}.}

\item{low_admin}{character string with the name of the
\code{low_df}. Must follow the format:
     \code{data_source_name} - \code{data_sheet_name} columns of
\code{context_AP}.}

\item{pop_df}{dataframe containing population figures for each admin unit
being analysed. This dataframe is used to weight the datasets when
aggregating from one administrative unit to another.
The dataframe must contain at a minimum:
\describe{
  \item{admin1}{contains names of administrative unit 1: regions,
  provinces, etc.}
  \item{admin2}{contains names of administrative unit 2: districts,
  departments, etc.}
  \item{total_pop}{contains the population figures associated with each
  administrative unit}
}
If additional administrative units are required, the name of the column
should follow this structure: admin+number (e.g. admin3).
If left to the default (NULL), the package will try to
download a dataframe from [HDX](https://data.humdata.org) Common
Operational Datasets (COD). This is a quite hazardous approach as HDX
COD's are not formatted in a consistent way.}

\item{context_AP}{data.frame with context specific analysis plan (AP) that
links the indicators in the WSC AP to the datasets used in the context
analysis.
See an example [here](https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing) or in \code{WSCprocessing::context_AP}.}
}
\value{
a data.frame with the aggregated data at a lower administrative level
}
\description{
As data is not always available at the required level, it might be needed to
assign/input data from a higher administrative unit to a lower one. For
instance,if a country has three administrative unit: region, province, and
communes, this function can help to bring data available only at the
region-level to a province-level data set.
}
\examples{
\dontrun{
assign_high2low(high_df = WSCprocessing::bfa_smart_2019_admin1,
     high_df_name = "smart_2019",
    context_AP = WSCprocessing::context_AP, WSC_AP = WSCprocessing::WSC_AP,
    LoAdmin_name = "admin2")
}
}
