% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfa_smart_2019_admin1.R
\docType{data}
\name{bfa_smart_2019_admin1}
\alias{bfa_smart_2019_admin1}
\title{Enquête Nutritionnelle Nationale SMART 2019}
\format{
A data frame with 12 rows and 12 variables:
}
\source{
\url{https://docs.google.com/spreadsheets/d/16noQMFiEifau26t7nV9q7NYQz79Xe5kZqiyedIQQCw0/edit?usp=sharing}
}
\usage{
bfa_smart_2019_admin1
}
\description{
A dataset with nutritional data for the whole of Burkina Faso. Extracted from \href{https://drive.google.com/file/d/18SPt_hRe7Ipd_QaWxJbm5g-qXKQ1YWrV/view?usp=sharing}{this report}
}
\keyword{datasets}
