% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{check_indicator}
\alias{check_indicator}
\title{Checks that indicator is in data}
\usage{
check_indicator(data, indicator)
}
\arguments{
\item{data}{data.frame}

\item{indicator}{character identifying a column to be checked}
}
\value{

}
\description{
Checks that indicator is in data
}
