% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{check_source}
\alias{check_source}
\title{Check that  list of indicators provided are in data}
\usage{
check_source(data, indicators_to_check)
}
\arguments{
\item{data}{data.frame}

\item{indicators_to_check}{vector of indicators to be checked}
}
\value{

}
\description{
Check that  list of indicators provided are in data
}
