% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_dataset}
\alias{clean_dataset}
\title{Clean dataset for missing data}
\usage{
clean_dataset(data, missings = NULL)
}
\arguments{
\item{data}{data.frame to be cleaned by replacing missing common missing
values with proper NA.}

\item{missings}{character list of possible missing data. By default includes:
'N/A','n/a',' ','(vide)','(empty)','d/m','','NA',and 'na'}
}
\value{

}
\description{
This function is a wrapper around \code{\link{rec_missing}} for
dataframes.
}
\examples{
df <- bfa_smart_2019_admin1
cleaned_df <- clean_dataset(df)
}
\seealso{
\code{\link{rec_missing}}

Other clean functions: 
\code{\link{normalise_string}()},
\code{\link{rec_missing}()}
}
\concept{clean functions}
