% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{get_dataAP}
\alias{get_dataAP}
\title{Get the analysis plan for a specific dataset.}
\usage{
get_dataAP(source_name, data_sheet_name, context_AP)
}
\arguments{
\item{source_name}{character string containing the name of the data source to
be analysed according to the analysis plan.}

\item{data_sheet_name}{character string of the sheet where the data is stored}

\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example \href{https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing}{here} or in \code{WSCprocessing::context_AP}.}
}
\value{
dataframe containing the analysis plan for the specified dataset.
}
\description{
This is essentially a formatting and filtering function.
}
\examples{

\dontrun{
this_dataAP <- get_dataAP("SMART-2019", "cleaned_data_admin1", context_AP)
}
}
