% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_population_dataframe.R
\name{get_pop_df}
\alias{get_pop_df}
\title{Get population data frame from HDX}
\usage{
get_pop_df(
  country_iso3,
  admin_level = NULL,
  google_api_key = NULL,
  pcodes_df = NULL
)
}
\arguments{
\item{country_iso3}{character string with the ISO3 country code for the
country analysed. See
{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{here} for more details.}

\item{admin_level}{character string identifying a specific administrative
level where the pcode should be looked for. Three values are possible:
"admin1", "admin2", and "admin3". For instance, if you are looking to
pcode the region of Bamako in Mali, you should specify "admin1", as there
is also an administrative unit 2 with the same name. By default the
lowest administrative unit found will be returned}

\item{google_api_key}{a string with a google API key. This is used to recode
administrative units. See vignette("googleway-vignette") for more details.}

\item{pcodes_df}{dataframe containing matching key between pcodes and admin
names. For more details see
\href{https://github.com/elliottmess/pcodesOCHA}{here}.}
}
\value{

}
\description{
Get population data frame from HDX
}
\examples{
\dontrun{
pop_df_example <- get_pop_df(country_iso3 = "BFA")
}
}
