% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{normalise_string}
\alias{normalise_string}
\title{Normalise string}
\usage{
normalise_string(string)
}
\arguments{
\item{string}{character string to be normalised. Accepts vectors too.}
}
\value{
normalise string
}
\description{
Normalise provided string by :
\enumerate{
\item removing extra spaces (\code{\link[stringr:str_trim]{stringr::str_squish()}}) and \code{\link[=trimws]{trimws()}}) 2. passing
the string to lower case (\code{\link[=tolower]{tolower()}}) 3. removing all accents
(\code{\link[stringi:stri_trans_general]{stringi::stri_trans_general()}}), with id = "Latin-ASCII") 4. removing other
special characters ('\link{-',.()/ }') and replacing with '_'
}
}
\examples{
normalise_string("àaF   kgfk")
}
\seealso{
Other clean functions: 
\code{\link{clean_dataset}()},
\code{\link{rec_missing}()}
}
\concept{clean functions}
