% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{rec_missing}
\alias{rec_missing}
\title{Recode missing values to NA}
\usage{
rec_missing(
  x,
  missings = c("N/A", "n/a", " ", "(vide)", "(empty)", "d/m", "", "NA", "na")
)
}
\arguments{
\item{x}{vector to be recoded.}

\item{missings}{character list of possible missing data. By default includes:
'N/A','n/a',' ','(vide)','(empty)','d/m','','NA',and 'na'}
}
\value{
recoded vector
}
\description{
Recode missing values to NA
}
\examples{
x <- c(1,2,NA, "N/A", "", "NA")
sum(is.na(x))
x_rec <- rec_missing(x)
sum(is.na(x_rec))
}
\seealso{
Other clean functions: 
\code{\link{clean_dataset}()},
\code{\link{normalise_string}()}
}
\concept{clean functions}
