% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{rename_vec}
\alias{rename_vec}
\title{Rename a vector according to the lists of parameters provided}
\usage{
rename_vec(vec, from, to)
}
\arguments{
\item{vec}{vector to be renamed}

\item{from}{list of values to look for}

\item{to}{list of values to change to}
}
\value{
vector renamed
}
\description{
Rename a vector according to the lists of parameters provided
}
\examples{
names(iris)
rename_vec(names(iris), from = "Sepal.Length", to = "Sepal length")

}
\seealso{
Other recode functions: 
\code{\link{recode_source}()},
\code{\link{recode_variable}()}
}
\concept{recode functions}
