% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{score_df_AP}
\alias{score_df_AP}
\title{Score dataset according to the Analysis Plan (AP) phases.}
\usage{
score_df_AP(
  data = NULL,
  data_name = NULL,
  data_type = NULL,
  agg_level,
  context,
  context_AP,
  WSC_AP
)
}
\arguments{
\item{data}{data.frame containing the data to be scored}

\item{data_name}{character string with the name of the \code{data} as is in the \code{data_source_name} column of \code{context_AP}.}

\item{data_type}{character string with the type of data source in \code{data}. Must be "area" or "hh".}

\item{agg_level}{character string specifying which column should be used to
aggregate the data. This is is typically an administrative unit (e.g. province,
region, departement, admin2, etc.)}

\item{context}{character string identifying the context to be used in the function call.
This is to be used if multiple context (geographical or temporal) are being
analysed. For instance, if data is used for Burkina Faso in 2020 and 2019,
this column can help distinguish the indicators.}

\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example [here](https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing) or in \code{WSCprocessing::context_AP}.}

\item{WSC_AP}{data.frame with the general WSC analysis plan (AP) than can be found \href{https://docs.google.com/spreadsheets/d/1TKxD_DyBTTN6onxYiooqtcI_TVSwPfeE-t7ZHK1zzMU/edit?usp=sharing}{here} or as an object in the package (```WSCprocessing::WSC_AP```)}
}
\value{
a data.frame containing the phase for each administrative level taken into consideration
}
\description{
This function scores datasets according to the phases documented in the AP under
the columns None/Minimal to Catastrophic.
}
\examples{
area_df <- score_df_AP(data = WSCprocessing::bfa_smart_2019_admin1, data_name = "smart_2019_admin1",
         data_type = "area",
         agg_level = "admin1", context = "bfa_2020", context_AP = WSCprocessing::context_AP,
         WSC_AP = WSCprocessing::WSC_AP)

hh_df <- score_df_AP(data = WSCprocessing::bfa_msna_2020, data_name = "msna_2020",
         data_type = "hh",
         agg_level = "admin1", context = "bfa_2020", context_AP = WSCprocessing::context_AP,
         WSC_AP = WSCprocessing::WSC_AP)

}
