% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_df.R
\name{score_source}
\alias{score_source}
\title{Score dataset according to the Analysis Plan (AP) phases.}
\usage{
score_source(data, data_AP, WSC_AP = WSCprocessing::WSC_AP, ...)
}
\arguments{
\item{data}{data.frame containing the data to be analysed.}

\item{data_AP}{data.frame with data specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used.
Most of those parameters are coming from context_AP and are automatically
passed from analyse_source_all_sheets.
The dataframe must contain:
\describe{
\item{data_sheet_name}{name of the sheet where the data is stored}
\item{data_worksheet_url}{URL of the worksheet to be used.}
\item{admin_level}{administrative level unit at which the data will be
aggregated}
\item{data_type}{type of data: either hh or area}
\item{context}{context identifying the time or place where the analysis
is conducted.}
\item{indicator_code_source}{list of indicators to be analysed as they
appear in the original dataset (data)}
\item{indicator_code}{list of indicators to be analysed as they
appear in the WSC analytical framework.}
}}

\item{WSC_AP}{data.frame with the general WSC analysis plan (AP) than can be found \href{https://docs.google.com/spreadsheets/d/1TKxD_DyBTTN6onxYiooqtcI_TVSwPfeE-t7ZHK1zzMU/edit?usp=sharing}{here} or as an object in the package (\code{WSCprocessing::WSC_AP})}

\item{...}{parameters to be passed to other functions, must be following the
format of a named list (e.g. WSC_AP = this_df).}
}
\value{
a data.frame containing the phase for each administrative level taken
into consideration
}
\description{
This function scores datasets according to the phases documented in the
\code{WSC_AP}  under the columns None/Minimal to Catastrophic.
}
\details{
This function differs from \code{\link[=analyse_source_all_sheets]{analyse_source_all_sheets()}} and \code{\link[=analyse_DAP]{analyse_DAP()}}
as it takes a unique data source rather than walking over multiple datasets.

#' The individual variables in the dataset are analysed with a weighted mean of
if it is a numerical  variable, or the weighted mean frequency of its factors
if it is a  character/factor variable.
}
\examples{
\dontrun{
}
}
\seealso{
Other score functions: 
\code{\link{score_WIS}()},
\code{\link{score_variable}()},
\code{\link{twenty_rule}()}
}
\concept{score functions}
