% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_df.R
\name{score_variable}
\alias{score_variable}
\title{Score individual variables}
\usage{
score_variable(var, survey_hh_data, agg_level)
}
\arguments{
\item{var}{character string identifying a variable present in \code{survey_hh_data}}

\item{survey_hh_data}{a srvyr::as_survey object containing necessary data and
survey information according to srvyr::as_survey definition.}

\item{agg_level}{character string specifying which column should be used to
aggregate the data. This is is typically an administrative unit (e.g. province,
region, departement, admin2, etc.)}
}
\value{
a data.frame containing the calculated scores
}
\description{
Takes an individual variable and scores it against the scales outlined in the
\code{WSC_AP} in columns "None/ minimal",	"Stressed", "Crisis",	"Critical",
and "Catastrophic".
}
\examples{
\dontrun{
 score_variable(var = "rsci_score", survey_hh_data = hh_data_scored,
 agg_level = "admin1")
}

}
\seealso{
Other score functions: 
\code{\link{score_WIS}()},
\code{\link{score_source}()},
\code{\link{twenty_rule}()}
}
\concept{score functions}
