% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{scoring_var}
\alias{scoring_var}
\title{Score individual variables}
\usage{
scoring_var(var, survey_hh_data, agg_level)
}
\arguments{
\item{var}{character string identifying a variable present in \code{survey_hh_data}}

\item{survey_hh_data}{a srvyr::as_survey object containing necessary data and
survey information according to srvyr::as_survey definition.}

\item{agg_level}{character string specifying which column should be used to
aggregate the data. This is is typically an administrative unit (e.g. province,
region, departement, admin2, etc.)}
}
\value{
a data.frame containing the calculated scores
}
\description{
Score individual variables
}
\examples{
\dontrun{
 scoring_var(var = "rsci_score", survey_hh_data = hh_data_scored, "admin1")
}

}
