% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring_WIS.R
\name{twenty_rule}
\alias{twenty_rule}
\title{Function to determine severity in a specific area based on the 20 percent rule.}
\usage{
twenty_rule(data, col_score, col_label, name_final_score, col_agg, col_value)
}
\arguments{
\item{data}{data.frame containing the data to be analysed.}

\item{col_score}{character string specifying the name the column with scores}

\item{col_label}{character string specifying the name the column with the
labels to be applied to the choices}

\item{name_final_score}{character string specifying the name score variable to
be filtered from col_score. The values of col_score for name_final_score
must be convertible to integers between 1 and 5.}

\item{col_agg}{character string specifying the name the column}

\item{col_value}{character string specifying which column should be used to
aggregate the data. This is is typically an administrative unit (e.g. province,
region, departement, admin2, etc.)}
}
\value{
a dataframe with the results of the 20 percent rule.
}
\description{
Function to determine severity in a specific area based on the 20 percent rule.
}
\details{
The 20 percent rule states that "an area is classified as the most severe phase that
    affects at least 20 percent of the population", based on the percentage of population
    in each of the five severity phases.

    The 20 percent rule is based on the \href{https://en.wikipedia.org/wiki/Pareto_principle}{Pareto principale}
}
\examples{
WSI_score_admin2 <- agg_score(data = WSCprocessing::bfa_msna_2020, context = "bfa_2020",
                              context_AP = WSCprocessing::context_AP,
                              WSC_AP = WSCprocessing::WSC_AP, agg_level = "admin2",
                              WIS_water = WSCprocessing::WIS_water, WIS_sanitation = WSCprocessing::WIS_sanitation,
                              WIS_final = WSCprocessing::WIS_final)

twenty_rule(data = WSI_score_admin2, col_score = "indicator",
            col_label = "choice", name_final_score = "score_final",
            col_agg = "admin2", col_value = "value")
}
