% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIS_sanitation.R
\docType{data}
\name{WIS_sanitation}
\alias{WIS_sanitation}
\title{Scoring reference matrix for the Sanitation component of the WASH Insecurity Score (WIS)}
\format{
A data frame with 54 rows and 5 variables:
\describe{
  \item{type_of_sanitation_facility}{Indicator code value for the type of sanitation facility used parameter}
  \item{sanitation_facility_sharing}{Indicator code value for the sharing of sanitation facility parameter}
  \item{access_to_soap}{Indicator code value for the access to soap parameter}
  \item{key_sanit}{Key to uniquely identify combinaison of type of sanitation facility, sharing of sanitation facility, and access to soap}
  \item{score_sanit}{Scoring value of individual key}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1UCr-G9gD6YZmiOHDoP95qiMkEqi9jMG3lfzzv7WCFnM/edit?usp=sharing}
}
\usage{
WIS_sanitation
}
\description{
A dataset containing the scoring reference matrix for the Sanitation component of the WASH Insecurity Score (WIS)
For more information on the WIS, please see the document \href{https://drive.google.com/file/d/1ZUapiDCMVTqlErdQnUjuErHFqdjpSOzr/view?usp=sharing}{here}.
}
\keyword{datasets}
