% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIS_water.R
\docType{data}
\name{WIS_water}
\alias{WIS_water}
\title{Scoring reference matrix for the Water component of the WASH Insecurity Score (WIS)}
\format{
A data frame with 82 rows and 5 variables:
\describe{
\item{Sufficiency of water}{Full description of the scale description used to attribute a score. For human control purpose}
\item{sufficiency_of_water}{Indicator code value for the sufficiency of water parameter}
\item{water_source_dist}{Indicator code value for the distance to water source}
\item{key_water}{Key to uniquely identify combinaison of variables sufficiency of water and distance to water source}
\item{score_water}{Scoring value of individual key}
}
}
\source{
\url{https://docs.google.com/spreadsheets/d/1UCr-G9gD6YZmiOHDoP95qiMkEqi9jMG3lfzzv7WCFnM/edit?usp=sharing}
}
\usage{
WIS_water
}
\description{
A dataset containing the scoring reference matrix for the Water component of the WASH Insecurity Score (WIS)
}
\keyword{datasets}
