% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse (conflicted).R, R/analyse.R
\name{analyse_country}
\alias{analyse_country}
\title{Analyse data for a specific country analysis plan at a specificied administative unit}
\usage{
analyse_country(
  context_AP,
  admin_analysis,
  country_iso3,
  pop_df = NULL,
  google_api_key = NULL,
  pcodes_df = NULL,
  ...
)

analyse_country(
  context_AP,
  admin_analysis,
  country_iso3,
  pop_df = NULL,
  google_api_key = NULL,
  pcodes_df = NULL,
  ...
)
}
\arguments{
\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example [here](https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing) or in ```WSCprocessing::context_AP```.}

\item{admin_analysis}{character string indicating at which administrative
unit the analysis should be aggregated.
Should follow the synthax: admin + admin_unit_number (e.g. admin1, admin2
, etc.)}

\item{country_iso3}{character string with the ISO3 country code for the
country analysed. See [here](https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3) for more details.}

\item{pop_df}{dataframe containing population figures for each admin unit
being analysed. This dataframe is used to weight the datasets when
aggregating from one administrative unit to another.
The dataframe must contain at a minimum:
\describe{
  \item{admin1}{contains names of administrative unit 1: regions,
  provinces, etc.}
  \item{admin2}{contains names of administrative unit 2: districts,
  departments, etc.}
  \item{total_pop}{contains the population figures associated with each
  administrative unit}
}
If additional administrative units are required, the name of the column
should follow this structure: admin+number (e.g. admin3).
If left to the default (NULL), the package will try to
download a dataframe from [HDX](https://data.humdata.org) Common
Operational Datasets (COD). This is a quite hazardous approach as HDX
COD's are not formatted in a consistent way.}

\item{google_api_key}{a string with a google API key. See vignette("googleway-vignette") for more details.}

\item{pcodes_df}{dataframe containing matching key between pcodes and admin
names. For more details see [here](https://github.com/elliottmess/pcodesOCHA).}

\item{...}{parameters to be passed to other functions, must be following the
format of a named list (e.g. WSC_AP = this_df).}
}
\value{



}
\description{
Analyse data for a specific country analysis plan at a specificied administative unit

Analyse data for a specific country analysis plan at a specificied administative unit
}
\examples{
\dontrun{
result <- analyse_country(context_AP, "admin2", "BFA", google_api_key = NULL)
}
assign_hiAdmin2loAdmin(HiAdmin_df = WSCprocessing::bfa_smart_2019_admin1,
    HiAdmin_name = "admin1", HiAdmin_df_name = "smart_2019",
    HiAdmin_df_sheet_name = "cleaned_data_admin1", context = "bfa_2020",
    context_AP = WSCprocessing::context_AP, WSC_AP = WSCprocessing::WSC_AP,
    LoAdmin_df = WSCprocessing::bfa_msna_2020, LoAdmin_name = "admin2")
\dontrun{
result <- analyse_country(context_AP, "admin2", "BFA", google_api_key = NULL)
}
}
