% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_DAP}
\alias{analyse_DAP}
\title{Analyse data according to the contextual data analysis plan (DAP)}
\usage{
analyse_DAP(
  context_AP,
  country_iso3,
  google_api_key = NULL,
  pcodes_df = NULL,
  ...
)
}
\arguments{
\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example \href{https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing}{here} or in \code{WSCprocessing::context_AP}.}

\item{country_iso3}{character string with the ISO3 country code for the
country analysed. See \href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{here} for more details.}

\item{google_api_key}{a string with a google API key. See
\code{vignette("googleway-vignette", package = "googleway")} for more details.}

\item{pcodes_df}{dataframe containing matching key between pcodes and admin
names. For more details see \href{https://github.com/elliottmess/pcodesOCHA}{here}.}

\item{...}{parameters to be passed to other functions, must be following the
format of a named list (e.g. WSC_AP = this_df).}
}
\value{
data.frame containing the analysed data
}
\description{
This function takes an analysis plan following the \code{\link{context_AP}}
structure, and a country iso3 code to conduct the analysis on it. The
function essentially walks through all the data sources listed in
\code{context_AP}, read dataset associated to it, sheet by sheet (for
googlesheets and excel files) or file by file (for csv or sav files). It then
cleans the data and find pcodes to have an uniform output that can be put in
a pipeline (such as \code{\link[=analyse_country]{analyse_country()}}).
}
\details{
The function outputs a data.frame.

The individual variables in the dataset are analysed with a weighted mean of
if it is a numerical  variable, or the weighted mean frequency of its factors
if it is a  character/factor variable.

Additional parameters are also recommended, especially \code{google_api_key},
which allows to match the administrative units with google places API. For
this, the \code{googleway} package is used. See
\code{vignette("googleway-vignette", package = "googleway")} for more details
}
\examples{
\dontrun{
context_AP <- WSCprocessing::context_AP

result <- analyse_DAP(context_AP, country_iso3 = "BFA")

}
}
\seealso{
Other analysis functions: 
\code{\link{analyse_country}()},
\code{\link{analyse_data}()},
\code{\link{analyse_source_all_sheets}()},
\code{\link{analyse_var}()}
}
\concept{analysis functions}
