% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse_source_all_sheets}
\alias{analyse_source_all_sheets}
\title{Analyse all data sheets in a data source}
\usage{
analyse_source_all_sheets(
  source_name,
  context_AP,
  WSC_AP = WSCprocessing::WSC_AP,
  ...
)
}
\arguments{
\item{source_name}{character string containing the name of the data source to
be analysed according to the analysis plan.}

\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example \href{https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing}{here} or in \code{WSCprocessing::context_AP}.}

\item{WSC_AP}{data.frame with the general WSC analysis plan (AP) than can be
found \href{https://docs.google.com/spreadsheets/d/1TKxD_DyBTTN6onxYiooqtcI_TVSwPfeE-t7ZHK1zzMU/edit?usp=sharing}{here} or as an object in the package (\code{WSCprocessing::WSC_AP})}

\item{...}{parameters to be passed to other functions, must be following the
format of a named list (e.g. WSC_AP = this_df).}
}
\value{
data.frame containing the analysed data
}
\description{
Function to analyse data source according to the analysis plan provided. This
function is similar to \code{\link[=analyse_DAP]{analyse_DAP()}} or \code{\link[=analyse_data]{analyse_data()}}, but walks over
all sheets individually in order to analyse each individually.
}
\examples{
\dontrun{
analysed_source <- analyse_source_all_sheets("REACH-MSNA-2020", WSCprocessing::context_AP)
}
}
\seealso{
Other analysis functions: 
\code{\link{analyse_DAP}()},
\code{\link{analyse_country}()},
\code{\link{analyse_data}()},
\code{\link{analyse_var}()}
}
\concept{analysis functions}
