% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_variable}
\alias{recode_variable}
\title{Recode variable according to the analysis plan}
\usage{
recode_variable(data, variable, data_AP)
}
\arguments{
\item{data}{data.frame containing the data to be recoded}

\item{variable}{string containing the name of the variable to be recoded. Must be present in \code{context_AP} in column \code{indicator_code}}

\item{data_AP}{data.frame with data specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used.
Most of those parameters are coming from context_AP and are automatically
passed from analyse_source_all_sheets.
The dataframe must contain:
\describe{
   \item{data_sheet_name}{name of the sheet where the data is stored}
   \item{data_worksheet_url}{URL of the worksheet to be used.}
   \item{admin_level}{administrative level unit at which the data will be
       aggregated}
   \item{data_type}{type of data: either hh or area}
   \item{context}{context identifying the time or place where the analysis
       is conducted.}
   \item{indicator_code_source}{list of indicators to be analysed as they
       appear in the original dataset (data)}
   \item{indicator_code}{list of indicators to be analysed as they
       appear in the WSC analytical framework.}
}}
}
\value{
data.frame with the variable recoded
}
\description{
Recode variable according to the analysis plan
}
\examples{
\dontrun{
var_recoded <- recode_variable(data = WSCprocessing::bfa_msna_2020,
variable = "critical_handwashing_times")
}
}
