% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{recode_variable}
\alias{recode_variable}
\title{Recode variable according to the analysis plan}
\usage{
recode_variable(data, variable, context_AP, data_source_name)
}
\arguments{
\item{data}{data.frame containing the data to be recoded}

\item{context_AP}{data.frame with context specific analysis plan (AP) that links
the indicators in the WSC AP to the datasets used in the context analysis.
See an example [here](https://docs.google.com/spreadsheets/d/1Pv1BBf32faE6J5tryubhVOsQJfGXaDb2t23KWGab52U/edit?usp=sharing) or in \code{WSCprocessing::context_AP}.}

\item{var}{string containing the name of the variable to be recoded. Must be present in \code{context_AP} in column \code{indicator_code}}
}
\value{
data.frame with the variable recoded
}
\description{
Recode variable according to the analysis plan
}
\examples{

var_recoded <- recode_variable(data = WSCprocessing::bfa_msna_2020,
variable = "critical_handwashing_times", context_AP = WSCprocessing::context_AP,
data_source_name = "bfa_msna_2020")

}
